+++ Sources/Heeler/Transport/EventsSession.swift
@@ -1213,6 +1213,9 @@ actor EventsSession {
         if case DeviceKeyStoreError.storedKeyCorrupt = error {
             return .deviceKeyCorrupt
         }
+        if case RSAKeyStoreError.storedKeyCorrupt = error {
+            return .rsaKeyCorrupt
+        }
         return .channelFailed(detail: String(describing: error))
     }